+++ spec/controllers/legal_panel_for_government/rm6360/admin/lot_data_controller_spec.rb
@@ -8,8 +8,8 @@ RSpec.describe LegalPanelForGovernment::RM6360::Admin::LotDataController do
   let(:supplier_framework_lot_services) { (1..5).map { |service_number| "RM6360.#{lot_number}.#{service_number}" }.map { |service_id| create(:supplier_framework_lot_service, supplier_framework_lot:, service_id:) } }
   let(:supplier_framework_lot_rates) { Position.where(lot_id: "RM6360.#{lot_number}", mandatory: true).pluck(:id).map { |position_id| create(:supplier_framework_lot_rate, supplier_framework_lot: supplier_framework_lot, jurisdiction: supplier_framework_lot_jurisdiction, position_id: position_id) } }
   let(:supplier_framework_lot_rates_non_gb) { Position.where(lot_id: "RM6360.#{lot_number}", mandatory: true).pluck(:id).map { |position_id| create(:supplier_framework_lot_rate, supplier_framework_lot: supplier_framework_lot, jurisdiction: supplier_framework_lot_jurisdiction_non_gb, position_id: position_id) } }
-  let(:supplier_framework_lot_jurisdiction) { create(:supplier_framework_lot_jurisdiction, supplier_framework_lot: supplier_framework_lot, jurisdiction_id: 'GB') }
-  let(:supplier_framework_lot_jurisdiction_non_gb) { create(:supplier_framework_lot_jurisdiction, supplier_framework_lot: supplier_framework_lot, jurisdiction_id: 'BM') }
+  let(:supplier_framework_lot_jurisdiction) { create(:supplier_framework_lot_jurisdiction, supplier_framework_lot: supplier_framework_lot, jurisdiction_id: 'RM6360.GB') }
+  let(:supplier_framework_lot_jurisdiction_non_gb) { create(:supplier_framework_lot_jurisdiction, supplier_framework_lot: supplier_framework_lot, jurisdiction_id: 'RM6360.BM') }
   let(:lot_number) { '1' }
 
   describe 'GET index' do
@@ -180,14 +180,14 @@ RSpec.describe LegalPanelForGovernment::RM6360::Admin::LotDataController do
         include_context 'when testing a section'
 
         it 'assigns jurisdictions' do
-          expect(assigns(:jurisdictions).pluck(:id)).to eq(['GB'])
+          expect(assigns(:jurisdictions).pluck(:id)).to eq(['RM6360.GB'])
         end
 
         it 'assigns supplier_framework_lot_rates' do
           assigned_supplier_framework_lot_rates = assigns(:supplier_framework_lot_rates)
 
           expect(assigned_supplier_framework_lot_rates.length).to eq(7)
-          expect(assigned_supplier_framework_lot_rates.map { |position_id, rates| [position_id, rates['GB'].id] }.sort).to eq(supplier_framework_lot_rates.map { |rate| [rate.position_id, rate.id] }.sort)
+          expect(assigned_supplier_framework_lot_rates.map { |position_id, rates| [position_id, rates['RM6360.GB'].id] }.sort).to eq(supplier_framework_lot_rates.map { |rate| [rate.position_id, rate.id] }.sort)
         end
 
         # rubocop:disable RSpec/NestedGroups
@@ -258,8 +258,8 @@ RSpec.describe LegalPanelForGovernment::RM6360::Admin::LotDataController do
           assigned_supplier_framework_lot_rates = assigns(:supplier_framework_lot_rates)
 
           expect(assigned_supplier_framework_lot_rates.length).to eq(8)
-          expect(assigned_supplier_framework_lot_rates.map { |position_id, rates| [position_id, rates['GB'].id] }.sort).to eq(supplier_framework_lot_rates.map { |rate| [rate.position_id, rate.id] }.sort)
-          expect(assigned_supplier_framework_lot_rates.map { |position_id, rates| [position_id, rates['BM'].id] }.sort).to eq(supplier_framework_lot_rates_non_gb.map { |rate| [rate.position_id, rate.id] }.sort)
+          expect(assigned_supplier_framework_lot_rates.map { |position_id, rates| [position_id, rates['RM6360.GB'].id] }.sort).to eq(supplier_framework_lot_rates.map { |rate| [rate.position_id, rate.id] }.sort)
+          expect(assigned_supplier_framework_lot_rates.map { |position_id, rates| [position_id, rates['RM6360.BM'].id] }.sort).to eq(supplier_framework_lot_rates_non_gb.map { |rate| [rate.position_id, rate.id] }.sort)
         end
 
         # rubocop:disable RSpec/NestedGroups
@@ -284,7 +284,7 @@ RSpec.describe LegalPanelForGovernment::RM6360::Admin::LotDataController do
         end
 
         it 'assigns supplier_framework_lot_jurisdiction_ids' do
-          expect(assigns(:supplier_framework_lot_jurisdiction_ids).sort).to eq(['BM', 'GB'])
+          expect(assigns(:supplier_framework_lot_jurisdiction_ids).sort).to eq(['RM6360.BM', 'RM6360.GB'])
         end
 
         # rubocop:disable RSpec/NestedGroups
@@ -417,7 +417,7 @@ RSpec.describe LegalPanelForGovernment::RM6360::Admin::LotDataController do
 
       context 'and the section is rates' do
         let(:section) { 'rates' }
-        let(:jurisdiction_id) { 'BM' }
+        let(:jurisdiction_id) { 'RM6360.BM' }
         let(:expected_template) { "legal_panel_for_government/rm6360/admin/lot_data/edit/_#{section}" }
 
         include_context 'when testing a section'
@@ -606,7 +606,7 @@ RSpec.describe LegalPanelForGovernment::RM6360::Admin::LotDataController do
           it 'creates a change log' do
             expect(change_log.change_type).to eq('update_supplier_framework_lot_rates')
             expect(change_log.change_data['id']).to eq(supplier_framework_lot.id)
-            expect(change_log.change_data['jurisdiction_id']).to eq('GB')
+            expect(change_log.change_data['jurisdiction_id']).to eq('RM6360.GB')
             expect(change_log.change_data['rates'].map { |rate_change| { position_id: rate_change['position_id'], before_rate: rate_change['before'].present?, after: rate_change['after'] } }).to eq(
               [
                 { after: 234567, before_rate: true, position_id: 'RM6360.1.1' },
@@ -760,7 +760,7 @@ RSpec.describe LegalPanelForGovernment::RM6360::Admin::LotDataController do
 
       context 'and the section is rates' do
         let(:section) { 'rates' }
-        let(:jurisdiction_id) { 'BM' }
+        let(:jurisdiction_id) { 'RM6360.BM' }
         let(:supplier_framework_lot_rates_non_gb) { (1..11).map { |position_number| create(:supplier_framework_lot_rate, supplier_framework_lot: supplier_framework_lot, jurisdiction: supplier_framework_lot_jurisdiction_non_gb, position_id: "RM6360.4a.#{position_number}") } }
         let(:rates) do
           {
@@ -829,7 +829,7 @@ RSpec.describe LegalPanelForGovernment::RM6360::Admin::LotDataController do
           it 'creates a change log' do
             expect(change_log.change_type).to eq('update_supplier_framework_lot_rates')
             expect(change_log.change_data['id']).to eq(supplier_framework_lot.id)
-            expect(change_log.change_data['jurisdiction_id']).to eq('BM')
+            expect(change_log.change_data['jurisdiction_id']).to eq('RM6360.BM')
             expect(change_log.change_data['rates'].map { |rate_change| { position_id: rate_change['position_id'], before_rate: rate_change['before'].present?, after: rate_change['after'] } }).to eq(
               [
                 { after: 123456, before_rate: true, position_id: 'RM6360.4a.1' },