+++ spec/controllers/supply_teachers/rm6376/admin/lot_data_controller_spec.rb
@@ -6,7 +6,7 @@ RSpec.describe SupplyTeachers::RM6376::Admin::LotDataController do
   let(:supplier_framework) { create(:supplier_framework, framework_id: 'RM6376') }
   let(:supplier_framework_lot) { create(:supplier_framework_lot, supplier_framework: supplier_framework, lot_id: "RM6376.#{lot_number}") }
   let(:supplier_framework_lot_rates) { Position.where(lot_id: "RM6376.#{lot_number}").pluck(:id).map { |position_id| create(:supplier_framework_lot_rate, supplier_framework_lot: supplier_framework_lot, jurisdiction: supplier_framework_lot_jurisdiction, position_id: position_id) } }
-  let(:supplier_framework_lot_jurisdiction) { create(:supplier_framework_lot_jurisdiction, supplier_framework_lot: supplier_framework_lot, jurisdiction_id: 'GB') }
+  let(:supplier_framework_lot_jurisdiction) { create(:supplier_framework_lot_jurisdiction, supplier_framework_lot: supplier_framework_lot, jurisdiction_id: 'RM6376.GB') }
   let(:supplier_framework_lot_branch) { create(:supplier_framework_lot_branch, supplier_framework_lot:) }
   let(:lot_number) { '1' }
 
@@ -384,7 +384,7 @@ RSpec.describe SupplyTeachers::RM6376::Admin::LotDataController do
           it 'creates a change log' do
             expect(change_log.change_type).to eq('update_supplier_framework_lot_rates')
             expect(change_log.change_data['id']).to eq(supplier_framework_lot.id)
-            expect(change_log.change_data['jurisdiction_id']).to eq('GB')
+            expect(change_log.change_data['jurisdiction_id']).to eq('RM6376.GB')
             expect(change_log.change_data['rates'].map { |rate_change| { position_id: rate_change['position_id'], before_rate: rate_change['before'].present?, after: rate_change['after'] } }).to eq(
               [
                 { after: 234567, before_rate: true, position_id: 'RM6376.1.1' },